Extract checkForUpdates to its own file
